feat: added user ownership verification on Project and Entry acess

+++ src/main/java/com/romm/timemanagement/controllers/EntryController.java
@@ -1,11 +1,8 @@
 package com.romm.timemanagement.controllers;
 
-import java.util.List;
-
-import com.romm.timemanagement.entities.Entry;
-import com.romm.timemanagement.services.EntryService;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,6 +12,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.romm.timemanagement.entities.Entry;
+import com.romm.timemanagement.services.EntryService;
+import com.romm.timemanagement.services.ProjectService;
+
 @RestController
 @RequestMapping(value="/entries")
 public class EntryController {
@@ -22,29 +23,43 @@ public class EntryController {
     @Autowired
     private EntryService entryService;
 
+    @Autowired
+    private ProjectService projectService;
+
     @PostMapping() // start
-    public Entry startEntry(@RequestBody Entry entry) {
-        return entryService.startEntry(entry);
+    public ResponseEntity<?> startEntry(@RequestBody Entry entry) {
+        if (projectService.isNotProjectOwner(entry.getProject().getId())) return ResponseEntity.status(HttpStatus.FORBIDDEN).body("User does not own the project");
+
+        return ResponseEntity.ok(entryService.startEntry(entry));
     }
 
     @PutMapping("/end/{id}") // end
-    public Entry endEntry(@PathVariable("id") Long id) {
-        return entryService.endEntry(id);
+    public ResponseEntity<?> endEntry(@PathVariable("id") Long id) {
+        if (projectService.isNotProjectOwner(entryService.findEntryById(id).getProject().getId())) return ResponseEntity.status(HttpStatus.FORBIDDEN).body("User does not own the project");
+
+        return ResponseEntity.ok(entryService.endEntry(id));
     }
 
     @DeleteMapping("/{id}") // delete
-    public void DeleteEntryById(@PathVariable("id") Long id) {
+    public ResponseEntity<?> DeleteEntryById(@PathVariable("id") Long id) {
+        if (projectService.isNotProjectOwner(entryService.findEntryById(id).getProject().getId())) return ResponseEntity.status(HttpStatus.FORBIDDEN).body("User does not own the project");
+
         entryService.deleteEntryById(id);
+        return ResponseEntity.ok().build();
     }
 
     @PutMapping("/{id}") // update
-    public Entry editEntry(@PathVariable("id") Long id, @RequestBody Entry entry) {
-        return entryService.editEntry(id, entry);
+    public ResponseEntity<?> editEntry(@PathVariable("id") Long id, @RequestBody Entry entry) {
+        if (projectService.isNotProjectOwner(entryService.findEntryById(id).getProject().getId())) return ResponseEntity.status(HttpStatus.FORBIDDEN).body("User does not own the project");
+
+        return ResponseEntity.ok(entryService.editEntry(id, entry));
     }
 
-    @GetMapping()
-    public List<Entry> findAll() {
-        return entryService.findAll();
+    @GetMapping("/{id}")
+    public ResponseEntity<?> findAllByProject(@PathVariable("id") Long id) {
+        if (projectService.isNotProjectOwner(entryService.findEntryById(id).getProject().getId())) return ResponseEntity.status(HttpStatus.FORBIDDEN).body("User does not own the project");
+
+        return ResponseEntity.ok(entryService.findAllByProjectId(id));
     }
 
 }
